api/requirements.txt
alembic==0.9.1
autoenv==1.0.0
bcrypt==3.1.3
cffi==1.10.0
click==6.7
coverage==4.0.3
Flask==0.12.1
Flask-API==0.7.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==1.0
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.4.2
python-coveralls==2.9.1
python-editor==1.0.3
PyYAML==3.12
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
shapely==1.6.2
